+++ src/app/(route)/service-introduction/_components/benefits/ServiceBenefitItem.tsx
